Added important cases in simplex algorithm.
Now if number of constraints in canonical LP problem is greater than number of variables,
then linear dependent constraints are omitted (corresponsibility of absolute form check is not implemented).
If number of linear independent constraints is equal to variables number,
then problem is solved as well defined system of linear equations.